Upgrade to gorm 2.0 and JWT new package complete.

+++ article/article.go
@@ -10,11 +10,11 @@ type Store interface {
 	CreateArticle(*model.Article) error
 	UpdateArticle(*model.Article, []string) error
 	DeleteArticle(*model.Article) error
-	List(offset, limit int) ([]model.Article, int, error)
-	ListByTag(tag string, offset, limit int) ([]model.Article, int, error)
-	ListByAuthor(username string, offset, limit int) ([]model.Article, int, error)
-	ListByWhoFavorited(username string, offset, limit int) ([]model.Article, int, error)
-	ListFeed(userID uint, offset, limit int) ([]model.Article, int, error)
+	List(offset, limit int) ([]model.Article, int64, error)
+	ListByTag(tag string, offset, limit int) ([]model.Article, int64, error)
+	ListByAuthor(username string, offset, limit int) ([]model.Article, int64, error)
+	ListByWhoFavorited(username string, offset, limit int) ([]model.Article, int64, error)
+	ListFeed(userID uint, offset, limit int) ([]model.Article, int64, error)
 	AddComment(*model.Article, *model.Comment) error
 	GetCommentsBySlug(string) ([]model.Comment, error)
 	GetCommentByID(uint) (*model.Comment, error)
